hospital_booking/requirements.txt
Django>=4.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0
